Rewrite the custom user provider to use eloquent
I had a moment of clarity.

Also rewrote the Traccar User model, now everything is alright in the
world.

Before this commit, I had hacked my way to the custom authentication and
user model, it was a weird mix of traccar's User Model with a sort of
DatabaseUserProvider, which returned a GenericUser. But when I wanted to
add relationships to the User Model, I couldn't, because it wasn't quite
a Model, but a Generic version of a model.

So I set sort of made it a model using inheritance, but it felt dirty,
so I stashed those changes (which actually worked), and when I was
thinking of what was going on, I realised I could use an Eloquent User
Provider, so I rewrote the Provider, and and customized the default User
Model (App\User) so it worked with my previous "Model".